Recommend that a defined "callback" function is still called on XHR Request fail
I would like to propose that a defined "callback" function is still called even when an XHR Request fails, or at the very least, the callback and callbackargs are passed as arguments to xhrRequestOnError to allow access to these.

I would expect my callback function to still run when receiving a "401" when attempting to authenticate a user. I don't consider this a fatal error and would therefore expect my callback function to execute, so that I can handle the "401"

The way I have implemented in the script allows developers to "hook" into xhrRequestOnError and do with it what they want and provide access to callback.